+++ src/components/MethodSelect/MethodSelect.stories.tsx
@@ -18,6 +18,10 @@ export default meta;
 type Story = StoryObj<typeof meta>;
 
 export const Default: Story = {
+  args: {
+    value: 'POST'
+  },
+
   render: (args) => {
     const [method, setMethod] = useState<HttpMethod>(args.value);
     return <MethodSelect value={method} onChange={setMethod} />;
@@ -26,7 +30,7 @@ export const Default: Story = {
 
 export const PostSelected: Story = {
   args: {
-    value: 'POST'
+    value: 'PATCH'
   },
   render: (args) => {
     const [method, setMethod] = useState<HttpMethod>(args.value);